requirements.txt
torch
torchvision
numpy
Pillow
scipy
scikit-image
matplotlib
tqdm
opencv-python-headless
